topic tree > Rename TopicTreeImpl to LocalTopicTree.
