+++ src/main/java/com/hivemq/mqtt/topic/tree/TopicTreeNode.java
@@ -31,7 +31,7 @@ class TopicTreeNode {
      * in each subscription and iterating the joint structure if the operation is relevant only to one kind of subscriptions.
      * <p>
      * The fields are NOT private to minimize the number of bypass methods.
-     * The class is intended to be used only as a part of {@link TopicTreeImpl}.
+     * The class is intended to be used only as a part of {@link LocalTopicTree}.
      */
     final @NotNull MatchingNodeSubscriptions wildcardSubscriptions;
     final @NotNull MatchingNodeSubscriptions exactSubscriptions;
